btm: Correct size of connection handle
It should be uint16_t rather than uint8_t.

Bug: 462902632
Bug: 452247122
Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.leaudio_broadcast_sink_feature

+++ system/bta/le_audio/client.cc
@@ -6231,7 +6231,7 @@ public:
                                                                    conn_handle);
   }
 
-  void IsoLinkQualityReadCb(uint8_t conn_handle, uint8_t cig_id, uint32_t tx_unacked_packets,
+  void IsoLinkQualityReadCb(uint16_t conn_handle, uint8_t cig_id, uint32_t tx_unacked_packets,
                             uint32_t tx_flushed_packets, uint32_t tx_last_subevent_packets,
                             uint32_t retransmitted_packets, uint32_t crc_error_packets,
                             uint32_t rx_unreceived_packets, uint32_t duplicate_packets) {
@@ -7155,7 +7155,7 @@ public:
     }
   }
 
-  void OnIsoLinkQualityRead(uint8_t conn_handle, uint8_t cig_id, uint32_t tx_unacked_packets,
+  void OnIsoLinkQualityRead(uint16_t conn_handle, uint8_t cig_id, uint32_t tx_unacked_packets,
                             uint32_t tx_flushed_packets, uint32_t tx_last_subevent_packets,
                             uint32_t retransmitted_packets, uint32_t crc_error_packets,
                             uint32_t rx_unreceived_packets, uint32_t duplicate_packets) {
